cleanup(core): move output prefix utilities to dedicated module
